Can now load NPZ and different 1-sample params When run from commandline, Vamb can now load TNF and RPKM directory from .npz files instead of calculating it. As a result of above, interface for specifying inputs on commandline has changed The parameters alpha and dropout now default to None. This means that it sets it optimally depending on whether there is one or more than one sample.

+++ vamb/__main__.py
@@ -25,66 +25,81 @@ import numpy as np
 import vamb
 
 ################################# DEFINE FUNCTIONS ##########################
-
 def log(string, logfile, indent=0):
     print(('\t' * indent) + string, file=logfile)
     logfile.flush()
 
-def calc_tnf(outdir, fastapath, mincontiglength, logfile):
+def calc_tnf(outdir, fastapath, tnfpath, namespath, lengthspath, mincontiglength, logfile):
     begintime = time.time()
+    log('\nLoading TNF', logfile, 0)
+    # If no path to FASTA is given, we load TNF from .npz files
+    if fastapath is None:
+        tnfs = vamb.vambtools.read_npz(tnfpath)
+        contignames = vamb.vambtools.read_npz(namespath)
+        contiglengths = vamb.vambtools.read_npz(lengthspath)
 
-    log('\nCalculating TNF', logfile, 0)
+        if not (len(tnfs) == len(contignames) == len(contiglengths)):
+            raise ValueError('Not all of TNFs, names and lengths are same length')
 
-    with open(fastapath, 'rb') as tnffile:
-        ret = vamb.parsecontigs.read_contigs(tnffile,
-                                             minlength=mincontiglength,
-                                             preallocate=True)
-    tnfs, contignames, contiglengths = ret
+        # Discard any sequence with a length below mincontiglength
+        mask = contiglengths >= mincontiglength
+        tnfs = tnfs[mask]
+        contignames = list(contignames[mask])
+        contiglengths = contiglengths[mask]
 
-    vamb.vambtools.write_npz(os.path.join(outdir, 'tnf.npz'), tnfs)
+    else:
+        with open(fastapath, 'rb') as tnffile:
+            ret = vamb.parsecontigs.read_contigs(tnffile,
+                                                 minlength=mincontiglength,
+                                                 preallocate=True)
+        tnfs, contignames, contiglengths = ret
+        vamb.vambtools.write_npz(os.path.join(outdir, 'tnf.npz'), tnfs)
 
     elapsed = round(time.time() - begintime, 2)
     ncontigs = len(contiglengths)
     nbases = contiglengths.sum()
-
-    log('Processed {} bases in {} sequences.'.format(nbases, ncontigs), logfile, 1)
-    log('Calculated TNF in {} seconds.'.format(elapsed), logfile, 1)
+    log('Kept {} bases in {} sequences.'.format(nbases, ncontigs), logfile, 1)
+    log('Processed TNF in {} seconds.'.format(elapsed), logfile, 1)
 
     return tnfs, contignames
 
-def calc_rpkm(outdir, bampaths, mincontiglength, minalignscore, subprocesses,
+def calc_rpkm(outdir, bampaths, rpkmpath, mincontiglength, minalignscore, subprocesses,
               ncontigs, logfile):
     begintime = time.time()
+    log('\nLoading RPKM', logfile)
+    # If bampaths is None, we load RPKM directly from .npz file
+    if bampaths is None:
+        rpkms = vamb.vambtools.read_npz(rpkmpath)
 
-    log('\nCalculating RPKM', logfile)
+    else:
+        log('Order of columns are:', logfile, 1)
+        log('\n\t'.join(bampaths), logfile, 1)
+        print('', file=logfile)
 
-    log('Order of columns are:', logfile, 1)
-    log('\n\t'.join(bampaths), logfile, 1)
-    print('', file=logfile)
+        log('Parsing {} BAM files with {} subprocesses.'.format(len(bampaths), subprocesses),
+           logfile, 1)
 
-    log('Parsing {} BAM files with {} subprocesses.'.format(len(bampaths), subprocesses),
-       logfile, 1)
-
-    dumpdirectory = os.path.join(outdir, 'tmp')
-    rpkms = vamb.parsebam.read_bamfiles(bampaths,
-                                        dumpdirectory=dumpdirectory,
-                                        minscore=minalignscore,
-                                        minlength=mincontiglength,
-                                        subprocesses=subprocesses,
-                                        logfile=logfile)
+        dumpdirectory = os.path.join(outdir, 'tmp')
+        rpkms = vamb.parsebam.read_bamfiles(bampaths,
+                                            dumpdirectory=dumpdirectory,
+                                            minscore=minalignscore,
+                                            minlength=mincontiglength,
+                                            subprocesses=subprocesses,
+                                            logfile=logfile)
+        print('', file=logfile)
 
     if len(rpkms) != ncontigs:
-        raise ValueError('Number of FASTA vs BAM file headers do not match. '
+        raise ValueError('Number of TNF and RPKM sequences do not match. '
                          'Are you sure the BAM files originate from same FASTA file '
                          'and have headers?')
 
-    vamb.vambtools.write_npz(os.path.join(outdir, 'rpkm.npz'), rpkms)
+    if bampaths is not None:
+        vamb.vambtools.write_npz(os.path.join(outdir, 'rpkm.npz'), rpkms)
+        shutil.rmtree(dumpdirectory)
+
     elapsed = round(time.time() - begintime, 2)
+    log('Processed RPKM in {} seconds.'.format(elapsed), logfile, 1)
 
-    print('', file=logfile)
-    log('Calculated RPKM in {} seconds.'.format(elapsed), logfile, 1)
-
-    shutil.rmtree(dumpdirectory)
     return rpkms
 
 def trainvae(outdir, rpkms, tnfs, nhiddens, nlatent, alpha, beta, dropout, cuda,
@@ -148,19 +163,23 @@ def cluster(outdir, latent, contignames, windowsize, minsuccesses, maxclusters,
     elapsed = round(time.time() - begintime, 2)
     log('Clustered contigs in {} seconds.'.format(elapsed), logfile, 1)
 
-def run(outdir, fastapath, bampaths, mincontiglength, minalignscore, subprocesses,
-         nhiddens, nlatent, nepochs, batchsize, cuda, alpha, beta, dropout, lrate,
-         batchsteps, windowsize, minsuccesses, minclustersize, maxclusters, logfile):
+def run(outdir, fastapath, tnfpath, namespath, lengthspath, bampaths, rpkmpath, mincontiglength,
+        minalignscore, subprocesses, nhiddens, nlatent, nepochs, batchsize,
+        cuda, alpha, beta, dropout, lrate, batchsteps, windowsize, minsuccesses,
+        minclustersize, maxclusters, logfile):
+
     log('Starting Vamb version ' + '.'.join(map(str, vamb.__version__)), logfile)
     log('Date and time is ' + str(datetime.datetime.now()), logfile, 1)
     begintime = time.time()
 
     # Get TNFs, save as npz
-    tnfs, contignames = calc_tnf(outdir, fastapath, mincontiglength, logfile)
+    tnfs, contignames = calc_tnf(outdir, fastapath, tnfpath, namespath, lengthspath,
+                                 mincontiglength, logfile)
 
     # Parse BAMs, save as npz
-    rpkms = calc_rpkm(outdir, bampaths, mincontiglength, minalignscore, subprocesses, len(contignames), logfile)
-
+    ncontigs = len(contignames)
+    rpkms = calc_rpkm(outdir, bampaths, rpkmpath, mincontiglength, minalignscore,
+                      subprocesses, ncontigs, logfile)
 
     # Train, save model
     mask, latent = trainvae(outdir, rpkms, tnfs, nhiddens, nlatent, alpha, beta,
@@ -180,9 +199,8 @@ def main():
     doc = """Vamb: Variational autoencoders for metagenomic binning.
 
     For advanced use and extensions of Vamb, check documentation of the package
-    at https://github.com/jakobnissen/vamb.
-    """
-    usage = "python runvamb.py OUTPATH FASTA BAMPATHS [OPTIONS ...]"
+    at https://github.com/jakobnissen/vamb."""
+    usage = "python runvamb.py OUTPATH FASTA|.NPZs BAMPATHS|RPKM [OPTIONS ...]"
     parser = argparse.ArgumentParser(
         description=doc,
         formatter_class=argparse.RawDescriptionHelpFormatter,
@@ -193,10 +211,20 @@ def main():
     helpos.add_argument('-h', '--help', help='print help and exit', action='help')
 
     # Positional arguments
-    reqos = parser.add_argument_group(title='Required arguments', description=None)
+    reqos = parser.add_argument_group(title='Output (required)', description=None)
     reqos.add_argument('outdir', help='output directory to create')
-    reqos.add_argument('fasta', help='path to fasta file')
-    reqos.add_argument('bamfiles', help='paths to (multiple) BAM files', nargs='+')
+
+    # TNF arguments
+    tnfos = parser.add_argument_group(title='TNF input (required)')
+    tnfos.add_argument('--fasta', metavar='', help='path to fasta file')
+    tnfos.add_argument('--tnfs', metavar='', help='path to .npz of TNF')
+    tnfos.add_argument('--names', metavar='', help='path to .npz of names of sequences')
+    tnfos.add_argument('--lengths', metavar='', help='path to .npz of seq lengths')
+
+    # RPKM arguments
+    rpkmos = parser.add_argument_group(title='RPKM input (required)')
+    rpkmos.add_argument('--bamfiles', metavar='', help='paths to (multiple) BAM files', nargs='+')
+    rpkmos.add_argument('--rpkm', metavar='', help='path to .npz of RPKM')
 
     # Optional arguments
     inputos = parser.add_argument_group(title='IO options', description=None)
@@ -217,11 +245,11 @@ def main():
     vaeos.add_argument('-l', dest='nlatent', metavar='', type=int,
                         default=40, help='latent neurons [40]')
     vaeos.add_argument('-a', dest='alpha',  metavar='',type=float,
-                        default=0.05, help='alpha, weight of TNF versus depth loss [0.05]')
+                        default=None, help='alpha, weight of TNF versus depth loss [Auto]')
     vaeos.add_argument('-b', dest='beta',  metavar='',type=float,
                         default=200.0, help='beta, capacity to learn [200.0]')
     vaeos.add_argument('-d', dest='dropout',  metavar='',type=float,
-                        default=0.2, help='dropout [0.2]')
+                        default=None, help='dropout [Auto]')
     vaeos.add_argument('--cuda', help='use GPU [False]', action='store_true')
 
     trainos = parser.add_argument_group(title='Training options', description=None)
@@ -251,7 +279,6 @@ def main():
         sys.exit()
 
     args = parser.parse_args()
-
     ######################### CHECK INPUT/OUTPUT FILES #####################
 
     # Outdir does not exist
@@ -264,14 +291,32 @@ def main():
     if parentdir and not os.path.isdir(parentdir):
         raise NotADirectoryError(parentdir)
 
-    # Contigs exists
-    if not os.path.isfile(args.fasta):
-        raise FileNotFoundError(args.fasta)
+    # Make sure only one TNF input is there
+    if args.fasta is None:
+        for path in args.tnfs, args.names, args.lengths:
+            if path is None:
+                raise argparse.ArgumentTypeError('Must specify either FASTA or the three .npz inputs')
+            if not os.path.isfile(path):
+                raise FileNotFoundError(path)
+    else:
+        for path in args.tnfs, args.names, args.lengths:
+            if path is not None:
+                raise argparse.ArgumentTypeError('Must specify either FASTA or the three .npz inputs')
+        if not os.path.isfile(args.fasta):
+            raise FileNotFoundError(args.fasta)
 
-    # All bamfiles exists
-    for bampath in args.bamfiles:
-        if not os.path.isfile(bampath):
-            raise FileNotFoundError(bampath)
+    # Make sure only one RPKM input is there
+    if args.bamfiles is None:
+        if args.rpkm is None:
+            raise argparse.ArgumentTypeError('Must specify either BAM files or RPKM input')
+        if not os.path.isfile(args.rpkm):
+            raise FileNotFoundError(args.rpkm)
+    else:
+        if args.rpkm is not None:
+            raise argparse.ArgumentTypeError('Must specify either BAM files or RPKM input')
+        for bampath in args.bamfiles:
+            if not os.path.isfile(bampath):
+                raise FileNotFoundError(bampath)
 
     ####################### CHECK ARGUMENTS FOR TNF AND BAMFILES ###########
     if args.minlength < 100:
@@ -287,13 +332,13 @@ def main():
     if args.nlatent < 1:
         raise argparse.ArgumentTypeError('Minimum 1 latent neuron, not {}'.format(args.latent))
 
-    if args.alpha <= 0 or args.alpha >= 1:
+    if args.alpha is not None and (args.alpha <= 0 or args.alpha >= 1):
         raise argparse.ArgumentTypeError('alpha must be above 0 and below 1')
 
     if args.beta <= 0:
         raise argparse.ArgumentTypeError('beta cannot be negative or zero')
 
-    if args.dropout < 0 or args.dropout >= 1:
+    if args.dropout is not None and (args.dropout < 0 or args.dropout >= 1):
         raise argparse.ArgumentTypeError('dropout must be in 0 <= d < 1.')
 
     if args.cuda and not torch.cuda.is_available():
@@ -327,29 +372,37 @@ def main():
 
     # This doesn't actually work, but maybe the PyTorch folks will fix it sometime.
     torch.set_num_threads(args.subprocesses)
-    subprocesses = min(DEFAULT_PROCESSES, len(args.bamfiles))
+    subprocesses = DEFAULT_PROCESSES
+    if args.bamfiles is not None:
+        subprocesses = min(DEFAULT_PROCESSES, len(args.bamfiles))
 
     ################### RUN PROGRAM #########################
     os.mkdir(args.outdir)
     logpath = os.path.join(args.outdir, 'log.txt')
 
     with open(logpath, 'w') as logfile:
-        run(args.outdir, args.fasta, args.bamfiles,
-             mincontiglength=args.minlength,
-             minalignscore=args.minascore,
-             subprocesses=subprocesses,
-             nhiddens=args.nhiddens,
-             nlatent=args.nlatent,
-             nepochs=args.nepochs,
-             batchsize=args.batchsize,
-             cuda=args.cuda,
-             alpha=args.alpha,
-             beta=args.beta,
-             dropout=args.dropout,
-             lrate=args.lrate,
-             batchsteps=args.batchsteps,
-             windowsize=args.windowsize,
-             minsuccesses=args.minsuccesses,
-             minclustersize=args.minsize,
-             maxclusters=args.maxclusters,
-             logfile=logfile)
+        run(args.outdir,
+            args.fasta,
+            args.tnfs,
+            args.names,
+            args.lengths,
+            args.bamfiles,
+            args.rpkm,
+            mincontiglength=args.minlength,
+            minalignscore=args.minascore,
+            subprocesses=subprocesses,
+            nhiddens=args.nhiddens,
+            nlatent=args.nlatent,
+            nepochs=args.nepochs,
+            batchsize=args.batchsize,
+            cuda=args.cuda,
+            alpha=args.alpha,
+            beta=args.beta,
+            dropout=args.dropout,
+            lrate=args.lrate,
+            batchsteps=args.batchsteps,
+            windowsize=args.windowsize,
+            minsuccesses=args.minsuccesses,
+            minclustersize=args.minsize,
+            maxclusters=args.maxclusters,
+            logfile=logfile)
